refactor: extract callback-safety primitives to callback.rs (Phase 1)
Move Group B (CbClass, ForceShutdown, ShutdownSeam + impls, NoShutdown,
PoisonFlag, report_contained_panic, guard_callback) and the callback_safety
test module out of lib.rs into a new callback.rs, re-exported at the crate root
for path stability. Pure move; no behavior/API change. 176 tests / 2 ignored.
